Code Improvements and Fixes
- Update the Serializers to inherit from the ValidateShelfSerializer when it is necessary.
	- Make the Container serializer inherit as well since the shelf is required for the queries and it is better this way than having sending the shelf without validation in the context or use more than one serializer without need.
	- Delete shelf and validation when necessary, so all validations and fields come from the ValidateShelfSerializer, avoid code repetition.
	- Update the shelfobject APIs to delete the shelf serializer creation and validation and then set shelf_id in the context.  That code is not required anymore becuause of the changes on the serializers.
- Update the shelfobject APIs to return the errors in the standard defined ({"errors": errors}) for those that didn't do it yet, it is better if everything follows conventions.
	- The manage shelfobject container form was not displaying the errors correctly for not following conventions.
	- Update the JS methods required (transfer in deny, add status and delete shelfobject) to display the errors correctly with the change in the API.
- Move the validate_measurement_unit_and_quantity method to utils, it is a better place for it.
- Update the unit tests required according to the changes performed.
- Some improvements in the utils code to avoid repetition and improve readability.
- Improve the CSS classes in the load errors method, that way when there is more than one error they are displayed correctly in the frontend and they look good.
- Other improvements and updates required found during the development of the changes.

+++ src/laboratory/api/shelfobject.py
@@ -40,8 +40,7 @@ from laboratory.shelfobject.serializers import IncreaseShelfObjectSerializer, \
     TransferOutShelfObjectSerializer, TransferObjectDataTableSerializer, \
     TransferInShelfObjectApproveWithContainerSerializer, ShelfObjectPk, \
     SearchShelfObjectSerializerMany, MoveShelfObjectWithContainerSerializer, \
-    ManageContainerSerializer, ValidateShelfSerializer, \
-    ValidateShelfInformationPositionSerializer
+    ManageContainerSerializer, ValidateShelfInformationPositionSerializer
 from laboratory.shelfobject.utils import save_increase_decrease_shelf_object, \
     move_shelfobject_partial_quantity_to, build_shelfobject_qr, \
     save_shelfobject_limits_from_serializer, \
@@ -466,16 +465,11 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
                                              "create_shelfobject")
         self.serializer_class = self._get_create_shelfobject_serializer(request, org_pk,
                                                                         lab_pk)
-        shelf_serializer = ValidateShelfSerializer(data=request.data,
-                                                   context={'laboratory_id':lab_pk})
-        errors = {}
-        shelf_serializer.is_valid(raise_exception=True)
-
         serializer = self.serializer_class['serializer'](data=request.data,
-                                                            context={"organization_id": org_pk,
-                                                                    "laboratory_id": lab_pk,
-                                                                    "shelf_id":shelf_serializer.validated_data['shelf'].pk})
+                                                         context={"organization_id": org_pk,
+                                                                  "laboratory_id": lab_pk})
         limit_serializer = ShelfObjectLimitsSerializer(data=request.data)
+        errors = {}
         if serializer.is_valid():
             if limit_serializer.is_valid(raise_exception=True):
                 shelfobject = self.serializer_class['method'](serializer,
@@ -511,8 +505,7 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         self._check_permission_on_laboratory(request, org_pk, lab_pk,
                                              "fill_increase_shelfobject")
         self.serializer_class = IncreaseShelfObjectSerializer
-        serializer = self.serializer_class(data=request.data, context={
-            "source_laboratory_id": self.laboratory.pk})
+        serializer = self.serializer_class(data=request.data, context={"source_laboratory_id": self.laboratory.pk})
         errors = {}
 
         if serializer.is_valid():
@@ -546,8 +539,7 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         self._check_permission_on_laboratory(request, org_pk, lab_pk,
                                              "fill_decrease_shelfobject")
         self.serializer_class = DecreaseShelfObjectSerializer
-        serializer = self.serializer_class(data=request.data, context={
-            "source_laboratory_id": self.laboratory.pk})
+        serializer = self.serializer_class(data=request.data, context={"source_laboratory_id": self.laboratory.pk})
         errors = {}
 
         if serializer.is_valid():
@@ -577,8 +569,7 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         """
         self._check_permission_on_laboratory(request, org_pk, lab_pk, "reserve")
         self.serializer_class = ReserveShelfObjectSerializer
-        serializer = self.serializer_class(data=request.data, context={
-            "source_laboratory_id": self.laboratory.pk})
+        serializer = self.serializer_class(data=request.data, context={"source_laboratory_id": self.laboratory.pk})
         errors = {}
         changed_data = ["laboratory", "organization", "user", "created_by"]
 
@@ -652,8 +643,7 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         """
         self._check_permission_on_laboratory(request, org_pk, lab_pk, "transfer_out")
         self.serializer_class = TransferOutShelfObjectSerializer
-        serializer = self.get_serializer(data=request.data,
-                                         context={"source_laboratory_id": lab_pk})
+        serializer = self.get_serializer(data=request.data, context={"source_laboratory_id": lab_pk})
         errors = {}
 
         if serializer.is_valid():
@@ -732,19 +722,16 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         """
         self._check_permission_on_laboratory(request, org_pk, lab_pk, "transfer_in_deny")
         self.serializer_class = TransferInShelfObjectSerializer
-        shelf_serializer = ValidateShelfSerializer(data=request.data,
-                                                   context={'laboratory_id':lab_pk})
+        serializer = self.get_serializer(data=request.data, context={"laboratory_id": lab_pk})
+        if serializer.is_valid():
+            utils.organilab_logentry(self.request.user,
+                                    serializer.validated_data['transfer_object'], DELETION,
+                                    relobj=self.laboratory)
+            serializer.validated_data['transfer_object'].delete()
+            return JsonResponse({'detail': _('The transfer in was denied successfully.')},
+                                status=status.HTTP_200_OK)
 
-        shelf_serializer.is_valid(raise_exception=True)
-        serializer = self.get_serializer(data=request.data, context={"laboratory_id": lab_pk,
-                                                                     "shelf_id":shelf_serializer.validated_data['shelf'].pk})
-        serializer.is_valid(raise_exception=True)
-        utils.organilab_logentry(self.request.user,
-                                 serializer.validated_data['transfer_object'], DELETION,
-                                 relobj=self.laboratory)
-        serializer.validated_data['transfer_object'].delete()
-        return JsonResponse({'detail': _('The transfer in was denied successfully.')},
-                            status=status.HTTP_200_OK)
+        return JsonResponse({"errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
     @action(detail=False, methods=['post'])
     def transfer_in_approve(self, request, org_pk, lab_pk, **kwargs):
@@ -759,12 +746,8 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         self._check_permission_on_laboratory(request, org_pk, lab_pk, "transfer_in_approve")
         transfer_obj = get_object_or_404(TranferObject, pk=request.data.get('transfer_object'))
         self.serializer_class = TransferInShelfObjectApproveWithContainerSerializer if transfer_obj.object.object.type == Object.REACTIVE else TransferInShelfObjectSerializer
-        shelf_serializer = ValidateShelfSerializer(data=request.data,
-                                                   context={'laboratory_id':lab_pk})
-        shelf_serializer.is_valid(raise_exception=True)
 
-        serializer = self.get_serializer(data=request.data, context={"laboratory_id": lab_pk, "organization_id": org_pk, "validate_for_approval": True,
-                                                                     "shelf_id":shelf_serializer.validated_data['shelf'].pk})
+        serializer = self.get_serializer(data=request.data, context={"laboratory_id": lab_pk, "organization_id": org_pk, "validate_for_approval": True})
 
         if serializer.is_valid():
             # once we get here everything is validated and ready for the transfer in to happen
@@ -810,24 +793,25 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         :return: JsonResponse with the status of the DELETE request
         """
         self._check_permission_on_laboratory(request, org_pk, lab_pk, "delete")
-        serializer = ShelfObjectDeleteSerializer(data=request.data, context={
-            "laboratory_id": self.laboratory.pk})
-        serializer.is_valid(raise_exception=True)
-        shelfobject = serializer.validated_data['shelfobj']
-        utils.organilab_logentry(self.request.user, shelfobject, DELETION,
-                                 relobj=self.laboratory)
-        log_object_change(request.user, lab_pk, shelfobject, shelfobject.quantity, 0,
-                          '', DELETION, _("Delete"))
+        serializer = ShelfObjectDeleteSerializer(data=request.data, context={"laboratory_id": self.laboratory.pk})
+        if serializer.is_valid():
+            shelfobject = serializer.validated_data['shelfobj']
+            utils.organilab_logentry(self.request.user, shelfobject, DELETION,
+                                    relobj=self.laboratory)
+            log_object_change(request.user, lab_pk, shelfobject, shelfobject.quantity, 0,
+                            '', DELETION, _("Delete"))
 
-        delete_container = serializer.validated_data.get('delete_container', False)
+            delete_container = serializer.validated_data.get('delete_container', False)
 
-        if shelfobject.container and delete_container:
-            shelfobject.container.delete()
+            if shelfobject.container and delete_container:
+                shelfobject.container.delete()
 
-        shelfobject.delete()
+            shelfobject.delete()
 
-        return JsonResponse({'detail': _('The item was deleted successfully')},
-                            status=200)
+            return JsonResponse({'detail': _('The item was deleted successfully')}, status=status.HTTP_200_OK)
+
+        return JsonResponse({"errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
+        
 
     @action(detail=True, methods=['post'])
     def create_comments(self, request, org_pk, lab_pk, pk, **kwargs):
@@ -950,18 +934,10 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
 
         self.serializer_class = MoveShelfObjectSerializer
         self.serializer_class_container = MoveShelfObjectWithContainerSerializer
-        shelf_serializer = ValidateShelfSerializer(data=request.data,
-                                                   context={'laboratory_id':lab_pk})
-        shelf_serializer.is_valid(raise_exception=True)
 
-        serializer = self.serializer_class(data=request.data, context={
-            "laboratory_id": lab_pk, "organization_id": org_pk,
-            "shelf_id":shelf_serializer.validated_data['shelf'].pk })
+        serializer = self.serializer_class(data=request.data, context={"laboratory_id": lab_pk, "organization_id": org_pk })
         serializer_container = self.serializer_class_container(data=request.data,
-                                                               context={
-                                                                   "laboratory_id": lab_pk,
-                                                                   "organization_id": org_pk,
-                                                               "shelf": shelf_serializer.validated_data['shelf'].pk})
+                                                               context={"laboratory_id": lab_pk, "organization_id": org_pk})
         errors = {}
 
         if serializer.is_valid():
@@ -1047,36 +1023,28 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
         :param lab_pk: laboratory related to reserved product and user permissions
         :param kwargs: extra params
         :return: save a status in it catalog, return success o error message
-
         """
 
         self._check_permission_on_laboratory(request, org_pk, lab_pk, "create_status")
 
         self.serializer_class = ShelfObjectStatusSerializer
         serializer = self.serializer_class(data=request.data)
-        if serializer.is_valid(raise_exception=True):
+        if serializer.is_valid():
             Catalog.objects.create(key='shelfobject_status',
                                    description=serializer.data['description'])
             return JsonResponse({'detail': _('The item was created successfully')},
                                 status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            
+        return JsonResponse({"errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
 
     @action(detail=False, methods=['post'])
     def manage_shelfobject_container(self, request, org_pk, lab_pk, **kwargs):
         self._check_permission_on_laboratory(request, org_pk, lab_pk, "manage_shelfobject_container")
         self.serializer_class=ManageContainerSerializer
-        shelf_serializer = ValidateShelfSerializer(data=request.data,
-                                                   context={'laboratory_id':lab_pk})
-        shelf_serializer.is_valid(raise_exception=True)
-        serializer = self.serializer_class(data=request.data, context={
-                'organization_id': org_pk,
-                'laboratory_id': lab_pk,
-                'shelf_id': shelf_serializer.validated_data['shelf'].pk
-            })
+        serializer = self.serializer_class(data=request.data, context={'organization_id': org_pk, 'laboratory_id': lab_pk})
 
-
-        if serializer.is_valid(raise_exception=True):
+        if serializer.is_valid():
             shelf_object = serializer.validated_data['shelf_object']
             #old_container = shelf_object.container
             new_container_object=get_or_create_container_based_on_selected_option(
@@ -1091,7 +1059,8 @@ class ShelfObjectViewSet(viewsets.GenericViewSet):
 
             return JsonResponse({'detail': _('The item was updated successfully')},
                                     status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            
+        return JsonResponse({"errors": serializer.errors}, status=status.HTTP_400_BAD_REQUEST)
 
 
 class SearchLabView(viewsets.GenericViewSet):
@@ -1242,8 +1211,7 @@ class SearchLabView(viewsets.GenericViewSet):
         self._check_permission_on_laboratory(request, org_pk, lab_pk)
         search_list, errors = {}, {}
         serializer = SearchShelfObjectSerializerMany(data=request.query_params,
-                                                     context={
-                                                         "source_laboratory_id": lab_pk})
+                                                     context={"source_laboratory_id": lab_pk})
 
         if serializer.is_valid():
             search_list = {
